backend/requirements.txt
asgiref==3.7.2
atomicwrites==1.4.1
attrs==23.1.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==2.0.12
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==41.0.1
defusedxml==0.7.1
Django==3.2
django-cors-headers==3.13.0
django-filter==22.1
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==4.7.2
djoser==2.1.0
drf-base64==2.0
gunicorn==20.1.0
idna==3.4
iniconfig==2.0.0
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.3
oauthlib==3.2.2
packaging==23.1
Pillow==9.5.0
pluggy==0.13.1
psycopg2-binary==2.9.3
py==1.11.0
pycparser==2.21
PyJWT==2.1.0
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3
requests==2.26.0
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.4.2
sqlparse==0.4.4
toml==0.10.2
typing_extensions==4.6.3
uritemplate==4.1.1
urllib3==1.26.16
django-extensions==3.2.3
django-import-export==3.2.0